% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{KinBNpackage-package}
\alias{KinBNpackage}
\alias{KinBNpackage-package}
\title{KinBNpackage: A package for kinship analysis}
\description{
KinBN is a free software for kinship analysis based on a Bayesian network. It can be applied to s hort tandem repeat (STR) markers commonly used in forensic genetics and calculates likelihood ratio.
}
